requirements.txt
streamlit
pandas
pyarrow
plotly
python-decouple